Moved spring-5-webclient example code to new test class, to verify everything is working as expected
